refactor:#50 컴포넌트 경로를 변경하고 새로운 파일로 이동
